List of GSZ-2

+++ cred_limit/gsz_save_item.php
@@ -14,9 +14,22 @@
 		case 'update':
 			$Brief_Name = mysql_escape_string( $_REQUEST['GSZ_Brief_Name'] );
 			$Full_Name = mysql_escape_string( $_REQUEST['GSZ_Full_Name'] );
+
+			//Проверяем, является ли параметр Id целым числом
+			if (!preg_match("/^\d+$/", $_REQUEST['Id']))
+			{
+				exit("Неверный формат URL-запроса");
+			}
+			
 			$query = 'UPDATE `GSZ` SET `Brief_Name`="'.$Brief_Name.'", `Full_Name`="'.$Full_Name.'" WHERE `Id`='.$_REQUEST['Id'];
 			break;
 		case 'delete':
+			//Проверяем, является ли параметр Id целым числом
+			if (!preg_match("/^\d+$/", $_REQUEST['Id']))
+			{
+				exit("Неверный формат URL-запроса");
+			}
+
 			$query = 'DELETE FROM `GSZ` WHERE `Id`='.$_REQUEST['Id'];
 			break;
 		default:
